[v0.23] + Анонимные трейды

+++ bin/callback_button_handler.py
@@ -8,9 +8,10 @@ from telegram.ext import ContextTypes
 
 from bin.achievements import bot_check_achievements
 from bin.anon_trade import generate_trade_keyboard, anon_trade_select_desired, \
-    anon_trade_confirm_sell, anon_trade_buy_card_show_offers, anon_trade_choose_menu, generate_trade_offers_keyboard
+    anon_trade_confirm_sell, anon_trade_buy_card_show_offers, anon_trade_choose_menu, generate_trade_offers_keyboard, \
+    anon_trade_show_my_offer, anon_trade_show_my_offers
 from bin.coinflip import coinflip_result
-from bin.collection import send_card_list, show_card, list_cards, get_collection_s
+from bin.collection import send_card_list, show_card, list_cards, get_collection_s, get_card_image
 from bin.market import market_sell_list_menu, shop_menu
 from lib.classes.user import User
 from lib.init import BOT_INFO
@@ -49,16 +50,16 @@ async def button_callback(update: Update, context: ContextTypes.DEFAULT_TYPE):
     if query.data.startswith("anon_trade_sell_page_"):
         page = int(query.data.split("_")[4])
         card_code = "c_" + re.search("anon_trade_sell_page_(.+)_c_(.+)", query.data).group(2)
-        context.user_data["page"] = page
-        mode = query.data.split("_")[2]
-        keyboard = generate_trade_keyboard(context, mode=mode, wts=card_code)
+        # context.user_data["page"] = page
+        mode: str = query.data.split("_")[2]
+        keyboard = generate_trade_keyboard(context, mode=mode, wts=card_code, page=page, user_id=query.from_user.id)
         await query.edit_message_reply_markup(reply_markup=keyboard)
 
     if query.data.startswith("anon_trade_buy_page_"):
         page = int(query.data.split("_")[4])
         context.user_data["page"] = page
         mode = query.data.split("_")[2]
-        keyboard = generate_trade_keyboard(context, mode=mode)
+        keyboard = generate_trade_keyboard(context, mode=mode, page=page, user_id=query.from_user.id)
         await query.edit_message_reply_markup(reply_markup=keyboard)
 
     if query.data.startswith("anon_trade_view_offer_page_"):
@@ -69,7 +70,23 @@ async def button_callback(update: Update, context: ContextTypes.DEFAULT_TYPE):
         except KeyError:
             await query.delete_message()
             return
-        keyboard = generate_trade_offers_keyboard(context, offers=context.user_data["anon_trade_offers"])
+        keyboard = generate_trade_offers_keyboard(context, offers=context.user_data["anon_trade_offers"],
+                                                  page=page, user_id=user.id)
+        if not keyboard:
+            await query.answer("Ошибка")
+            await query.delete_message()
+            return
+        await query.edit_message_reply_markup(reply_markup=keyboard)
+
+    if query.data.startswith("anon_trade_view_my_offers_page_"):
+        page = int(query.data.split("_")[-1])
+        context.user_data["anon_trade_my_offers_page"] = page
+        try:
+            keyboard = generate_trade_offers_keyboard(context, type="my_offers",
+                                                      offers=context.user_data["anon_trade_my_offers"], page=page)
+        except KeyError:
+            await query.delete_message()
+            return
         if not keyboard:
             await query.answer("Ошибка")
             await query.delete_message()
@@ -87,7 +104,7 @@ async def button_callback(update: Update, context: ContextTypes.DEFAULT_TYPE):
         page = int(inline_keyboard[-1][page_index]['text']) - 1
         await show_card(query, context, in_market=False, page=page)
 
-    elif query.data.startswith("close_"):
+    elif query.data.startswith("close_") or query.data == "anon_trade_my_offers_close":
         if "card" in query.data:
             context.user_data['page'] = query.data.split("_")[-1]
             context.user_data["closed_card"] = True
@@ -370,7 +387,8 @@ async def button_callback(update: Update, context: ContextTypes.DEFAULT_TYPE):
                                          message_id=query.message.message_id)
 
         await context.bot.send_message(user.id,
-                                       text="Поздравляем! Лента получена и отображается в вашем, а карточки списаны. "
+                                       text="Поздравляем! Лента получена и отображается в вашем профиле, "
+                                            "а карточки списаны. "
                                             "Может, ещё одну?")
         await query.answer()
 
@@ -379,17 +397,17 @@ async def button_callback(update: Update, context: ContextTypes.DEFAULT_TYPE):
         await query.delete_message()
         card_code = "c_" + re.search("anon_trade_add_c_(.+)_(.+)", query.data).group(1)
         page = re.search("anon_trade_add_c_(.+)_(.+)", query.data).group(2)
-        context.user_data["wts"] = card_code
-        context.user_data["page"] = page
-        await anon_trade_select_desired(update, context)
+        # context.user_data["wts"] = card_code
+        # context.user_data["page"] = page
+        await anon_trade_select_desired(update, context, wts=card_code, page=page)
 
     elif query.data.startswith("anon_trade_sell_c_"):
         await query.answer()
         wts = "c_" + re.search("anon_trade_sell_c_(.+)_c_(.+)", query.data).group(2)
         wtb = "c_" + re.search("anon_trade_sell_c_(.+)_c_(.+)", query.data).group(1)
-        context.user_data["wtb"] = wtb
-        context.user_data["wts"] = wts
-        await anon_trade_confirm_sell(update, context)
+        # context.user_data["wtb"] = wtb
+        # context.user_data["wts"] = wts
+        await anon_trade_confirm_sell(update, context, wts=wts, wtb=wtb)
 
     elif query.data.startswith("anon_trade_confirm_sell_"):
         await query.answer()
@@ -407,24 +425,36 @@ async def button_callback(update: Update, context: ContextTypes.DEFAULT_TYPE):
 
         await list_cards(update, context)
 
-    elif query.data == "anon_trade_cancel_sell":
+    elif query.data.startswith("anon_trade_cancel_sell_"):
+        card_code = "c_" + query.data.split("_")[-1]
         await query.answer("Отменено")
         await query.delete_message()
+        await show_card(query, context, in_market=False, card_code=card_code)
 
     elif query.data.startswith("anon_trade_close_sell_"):
         card_code = re.search("anon_trade_close_sell_(.+)", query.data).group(1)
-        context.user_data["card_code"] = card_code
-        await show_card(query, context, in_market=False)
+        # context.user_data["card_code"] = card_code
+        await show_card(query, context, in_market=False, card_code=card_code)
 
     elif query.data.startswith("anon_trade_buy_c_"):
         card_code = "c_" + re.search("anon_trade_buy_c_(.+)", query.data).group(1)
-        context.user_data["card_code"] = card_code
-        await anon_trade_buy_card_show_offers(update, context)
+        # context.user_data["card_code"] = card_code
+        await anon_trade_buy_card_show_offers(update, context, card_code=card_code)
 
     elif query.data.startswith("anon_trade_view_buy_list"):
         await query.answer()
-        context.user_data["page"] = 0
-        await anon_trade_choose_menu(update, context)
+        try:
+            page = context.user_data['anon_trade_page']
+        except KeyError:
+            context.user_data['anon_trade_page'] = 0
+            page = 0
+        # context.user_data["page"] = 0
+        await anon_trade_choose_menu(update, context, page=page, user_id=query.from_user.id)
+
+    elif query.data.startswith("anon_trade_close_buy"):
+        await query.answer()
+        # context.user_data["page"] = 0
+        await query.delete_message()
 
     elif query.data.startswith("anon_trade_view_offer_"):
         await query.answer()
@@ -438,7 +468,8 @@ async def button_callback(update: Update, context: ContextTypes.DEFAULT_TYPE):
 
         resp = f"Вы согласны отдать карту {cards_dict[wtb]['name']} и получить карту {cards_dict[wts]['name']}?"
         keyboard = InlineKeyboardMarkup([[InlineKeyboardButton("Да",
-                                                               callback_data=f"anon_trade_offer_confirm_{receiver_id}_{wts}_{wtb}")],
+                                                               callback_data=f"anon_trade_offer_confirm_{receiver_id}_"
+                                                                             f"{wts}_{wtb}")],
                                          [InlineKeyboardButton("Нет",
                                                                callback_data=f"anon_trade_offer_reject_{wts}")]])
 
@@ -446,14 +477,17 @@ async def button_callback(update: Update, context: ContextTypes.DEFAULT_TYPE):
                                       reply_markup=keyboard)
 
     elif query.data.startswith("anon_trade_offer_confirm_"):
-        receiver_id, wts, wtb = re.search("anon_trade_offer_confirm_(.+)_c_(.+)_c_(.+)", query.data).groups()
+        receiver_id, wtb, wts = re.search("anon_trade_offer_confirm_(.+)_c_(.+)_c_(.+)", query.data).groups()
         wts = "c_" + wts
         wtb = "c_" + wtb
 
         receiver = User.get(None, int(receiver_id))
         try:
             user.collection.remove(wts)
-            receiver.collection.append(wtb)
+            user.collection.append(wtb)
+            receiver.anon_trade.remove({'wts': wtb, 'wtb': wts})
+            receiver.collection.append(wts)
+            await anon_trade_choose_menu(update, context)
         except ValueError:
             await query.delete_message()
             return
@@ -461,11 +495,70 @@ async def button_callback(update: Update, context: ContextTypes.DEFAULT_TYPE):
         user.write()
         receiver.write()
 
-        await context.bot.send_message(user.id,
-                                       f"Успешно!\n\nПолучена карта {cards_dict[wtb]['name']}.")
+        wtb_limited, wts_limited = wtb.startswith("c_9"), wts.startswith("c_9")
 
-        await context.bot.send_message(receiver.id,
-                                       f"Кто-то обменялся в вами картой!\n\nПолучена карта {cards_dict[wts]['name']}.")
+        wtb_card_pic_id = get_card_image(wtb, is_limited=wtb_limited)
+        wts_card_pic_id = get_card_image(wts, is_limited=wts_limited)
 
-        await query.delete_message()
+        # Отправляем карту wtb
+        if wtb_limited:
+            await context.bot.send_animation(user.id,
+                                             animation=wtb_card_pic_id,
+                                             caption=f"Успешно!\nПолучена карта: {cards_dict[wtb]['name']}.")
+        else:
+            await context.bot.send_photo(user.id,
+                                         photo=wtb_card_pic_id,
+                                         caption=f"Успешно!\nПолучена карта: {cards_dict[wtb]['name']}.")
+
+        # Отправляем карту wts
+        if wts_limited:
+            await context.bot.send_animation(receiver.id,
+                                             animation=wts_card_pic_id,
+                                             caption=f"Кто-то обменялся с вами картой!\nПолучена карта: "
+                                                     f"{cards_dict[wts]['name']}.")
+        else:
+            await context.bot.send_photo(receiver.id,
+                                         photo=wts_card_pic_id,
+                                         caption=f"Кто-то обменялся с вами картой!\nПолучена карта: "
+                                                 f"{cards_dict[wts]['name']}.")
+
+        # await query.delete_message()
         await query.answer()
+
+    elif query.data.startswith("anon_trade_my_offer_c_"):
+        wts, wtb = re.search("anon_trade_my_offer_c_(.+)_c_(.+)", query.data).groups()
+        wts = "c_" + wts
+        wtb = "c_" + wtb
+
+        await anon_trade_show_my_offer(update, context, wts=wts, wtb=wtb)
+
+    elif query.data.startswith("anon_trade_show_my_offers"):
+        page = context.user_data.get("anon_trade_my_offers_page", 0)
+        await anon_trade_show_my_offers(update, context, page=page)
+
+    elif query.data.startswith("anon_trade_my_offer_remove_c_"):
+        wts, wtb = re.search("anon_trade_my_offer_remove_c_(.+)_c_(.+)", query.data).groups()
+        resp = "Вы уверены, что хотите отменить предложение?"
+        keyboard = InlineKeyboardMarkup([[InlineKeyboardButton("Да",
+                                                               callback_data="anon_trade_my_offer_"
+                                                                             f"remove_confirm_c_{wts}_c_{wtb}")],
+                                         [InlineKeyboardButton("Нет",
+                                                               callback_data="anon_trade_show_my_offers")]])
+        await query.edit_message_text(resp,
+                                      reply_markup=keyboard)
+
+    elif query.data.startswith("anon_trade_my_offer_remove_confirm_c_"):
+        page = context.user_data.get("page", 0)
+        wts, wtb = re.search("anon_trade_my_offer_remove_confirm_c_(.+)_c_(.+)", query.data).groups()
+        offer = {'wts': f'c_{wts}', 'wtb': f"c_{wtb}"}
+        try:
+            user.anon_trade.remove(offer)
+            user.collection.append("c_"+wts)
+            user.write()
+        except ValueError:
+            await query.answer("Предложения не существует!")
+            await anon_trade_show_my_offers(update, context, page=page)
+            return
+
+        await anon_trade_show_my_offers(update, context, page=page)
+        await query.answer("Предложение отменено успешно!")
